after code refactor

+++ src/main/java/com/cfm/tiri/repositories/TrailerRepository.java
@@ -6,12 +6,12 @@ import org.springframework.data.repository.query.Param;
 
 import com.cfm.tiri.domain.Trailer;
 
-public interface TrailerRepository extends CrudRepository<Trailer, Integer> {
+public interface TrailerRepository extends CrudRepository<Trailer, Long> {
 
 	Iterable<Trailer> findByActive(boolean active);
 	// Zgodnie z
 	// https://docs.spring.io/spring-data/data-commons/docs/1.6.1.RELEASE/reference/html/repositories.html
 
 	@Query(value = "SELECT * FROM trailer t LEFT JOIN (SELECT * FROM squad where active = 1 AND id_squad not in (:squadId)) s on s.id_trailer=t.id_trailer WHERE t.active = 1 and s.id_trailer is null", nativeQuery = true)
-	Iterable<Trailer> findActiveAndFree(@Param("squadId") int squadId);
+	Iterable<Trailer> findActiveAndFree(@Param("squadId") long squadId);
 }
